distributor: Drop flag for non-direct OTLP translation
