+++ docs/sources/mimir/configure/about-versioning.md
@@ -82,8 +82,6 @@ The following features are currently experimental:
     - `-distributor.max-otlp-request-size`
   - Enforce a maximum pool buffer size for write requests
     - `-distributor.max-request-pool-buffer-size`
-  - Enable direct translation from OTLP write requests to Mimir equivalents
-    - `-distributor.direct-otlp-translation-enabled`
   - Enable conversion of OTel start timestamps to Prometheus zero samples to mark series start
     - `-distributor.otel-created-timestamp-zero-ingestion-enabled`
 - Hash ring
